Move leaflet stuff to a map/ subdirectory
