+++ frontend/src/layers/BoundaryLayerDepth.tsx
@@ -2,7 +2,7 @@ import { Forecast, ForecastPoint } from "../data/Forecast";
 import * as L from 'leaflet';
 import { ColorScale, Color } from "../ColorScale";
 import { JSX } from "solid-js";
-import { DataSource } from "../CanvasLayer";
+import { DataSource } from "../map/CanvasLayer";
 
 export class BoundaryLayerDepth implements DataSource {
