+++ frontend/src/layers/ThQ.tsx
@@ -3,7 +3,7 @@ import { ColorScale, Color } from "../ColorScale";
 import * as L from 'leaflet';
 import { drawWindArrow } from "../shapes";
 import { windColor } from "./Wind";
-import { DataSource } from "../CanvasLayer";
+import { DataSource } from "../map/CanvasLayer";
 import { JSX } from "solid-js";
 
 export const colorScale = new ColorScale([
